Finish assessment batches 1-4 and tenant EA discount for executive ROI.
Ship identity-provider diagnostics, reliability guardrails, AI observability, extractor validation, graph cache sizing, and EA discount percentage on tenant cost settings with TenantEaDiscountMath. Rescore LATEST.md to 94.00%.

+++ ArchLucid.TestSupport/SqlExceptionTestFactory.cs
@@ -29,7 +29,7 @@ public static class SqlExceptionTestFactory
                 "message",
                 "procedure",
                 0,
-                0,
+                (uint)0,
                 null
             ],
             null)!;
